Move formatting functions logic to TagLinker
