+++ .rubocop_todo.yml
@@ -31,10 +31,10 @@ Lint/AssignmentInCondition:
     - 'app/controllers/traces_controller.rb'
     - 'app/controllers/users_controller.rb'
     - 'app/helpers/application_helper.rb'
-    - 'app/helpers/browse_tags_helper.rb'
     - 'app/mailers/user_mailer.rb'
     - 'lib/nominatim.rb'
     - 'lib/osm.rb'
+    - 'lib/tag_linker.rb'
     - 'script/deliver-message'
 
 # Offense count: 858
