+++ app/views/browse/_tag.html.erb
@@ -2,5 +2,5 @@
 
 <tr>
   <th class="py-1 border-secondary-subtle table-secondary fw-normal" dir="auto"><%= format_key(tag[0]) %></th>
-  <td class="py-1 border-secondary-subtle border-start" dir="auto"><%= format_value(tag[0], tag[1]) %></td>
+  <td class="py-1 border-secondary-subtle border-start" dir="auto"><%= format_value(tag[0], tag[1], :html_only => true) %></td>
 </tr>
